Pinewood.Customers/Pinewood.Customers.Repo/CustomerRepository.cs
using Microsoft.EntityFrameworkCore;
using Pinewood.Customers.Db;
using Pinewood.Customers.Models.DbModels;
using Pinewood.Customers.Models.Interfaces;

public class CustomerRepository : IRepository<Customer>
{
    private readonly PinewoodCustomersDbContext _context;

    public CustomerRepository(PinewoodCustomersDbContext context)
    {
        _context = context;
    }

    public async Task<Customer> GetByIdAsync(int id)
    {
        return await _context.Customers.FindAsync(id);
    }

    public async Task<List<Customer>> GetAllAsync()
    {
        return await _context.Customers.ToListAsync();
    }

    public async Task<Customer> AddAsync(Customer customer)
    {
        _context.Customers.Add(customer);
        await _context.SaveChangesAsync();
        return customer;
    }

    public async Task<Customer> UpdateAsync(Customer customer)
    {
        _context.Entry(customer).State = EntityState.Modified;
        await _context.SaveChangesAsync();
        return customer;
    }

    public async Task DeleteAsync(int id)
    {
        var customer = await GetByIdAsync(id);
        if (customer != null)
        {
            _context.Customers.Remove(customer);
            await _context.SaveChangesAsync();
        }
    }
    public async Task SaveChangesAsync()
    {
        await _context.SaveChangesAsync();
    }
}
